+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.6.18
 
 * Some monsters, e.g. Abhorrent Overlord could import with the spellcaster level set incorrectly.
+* Monsters: Sephek Kaltro did not import Misty Step. (Odd formating)
+* Monsters: Aboleth tentacle damage was not correct. Acid is now split to the Other formula field.
 
 # 2.6.17